Refactor App.tsx to modular architecture (under 100 lines per file)
Broke down 777-line App.tsx into focused, maintainable modules:

**Hooks:**
- useAppState.ts (56 lines) - Centralized state management
- useKeyboardShortcuts.ts (77 lines) - Global keyboard shortcuts

**Handlers:**
- commandHandlers.ts (43 lines) - Embed and note command handlers
- commandProcessor.ts (87 lines) - Command routing and processing
- messageHandlers.ts (99 lines) - Message execution and UI handlers

**Services:**
- toolPlanning.ts (124 lines) - AI-powered tool and visualization planning
- toolGeneration.ts (92 lines) - Tool and visualization generation
- visualizationPrompt.ts (100 lines) - Visualization prompt builder

**Utils:**
- commandDetector.ts (18 lines) - Slash command detection

**Components:**
- MaximizedEmbedModal.tsx (55 lines) - Extracted modal component
- App.tsx (76 lines) - Streamlined main component

Benefits:
- Single Responsibility Principle - each file has one clear purpose
- Improved testability and maintainability
- Better code organization and discoverability
- Reduced cognitive load when reading code
- Easier to locate and fix bugs
